requirements.txt
Pillow
numpy
opencv-python
scikit-image
networkx
